backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.35.2
sentence-transformers==2.7.0
weaviate-client==4.4.0
pydantic==2.5.0
torch==2.1.1
lxml==4.9.3
huggingface-hub==0.23.0
